backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
spotipy==2.24.0
SQLAlchemy==2.0.31
requests==2.32.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
